+++ src/pytask_parallel/config.py
@@ -34,5 +34,5 @@ def pytask_parse_config(config: dict[str, Any]) -> None:
 @hookimpl
 def pytask_post_parse(config: dict[str, Any]) -> None:
     """Disable parallelization if debugging is enabled."""
-    if config["pdb"] or config["trace"]:
+    if config["pdb"] or config["trace"] or config["dry_run"]:
         config["n_workers"] = 1