+++ CHANGES.md
@@ -8,6 +8,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask-parallel) and
 ## 0.4.0 - 2023-xx-xx
 
 - {pull}`62` deprecates Python 3.7.
+- {pull}`64` aligns pytask-parallel with pytask v0.4.0rc2.
 
 ## 0.3.1 - 2023-05-27