+++ CHANGES.md
@@ -10,6 +10,8 @@ releases are available on [PyPI](https://pypi.org/project/pytask-parallel) and
 - {pull}`62` deprecates Python 3.7.
 - {pull}`64` aligns pytask-parallel with pytask v0.4.0rc2.
 - {pull}`66` deactivates parallelization for dry-runs.
+- {pull}`67` fixes parallelization with partialed task functions.
+- {pull}`68` updates dependencies and syntaxes.
 
 ## 0.3.1 - 2023-05-27